Finished routing + separated routes in express.router
tested everything, works! :)
